# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Quản Lý Cân Xe API",
    "version": "1.0.0",
    "description": "API tài liệu cho hệ thống quản lý cân xe.\n\nHệ thống này cung cấp API để quản lý việc cân xe, thông tin khách hàng, \nhàng hóa và các chức năng liên quan đến quản lý phiếu cân.\n",
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Your Name",
      "url": "https://yourwebsite.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Khachhang": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "id": {
            "type": "number",
            "description": "ID tự động tạo của khách hàng"
          },
          "ma": {
            "type": "string",
            "description": "Mã khách hàng"
          },
          "ten": {
            "type": "string",
            "description": "Tên khách hàng"
          },
          "diachi": {
            "type": "string",
            "description": "Địa chỉ"
          },
          "dienthoai": {
            "type": "string",
            "description": "Số điện thoại"
          }
        },
        "example": {
          "id": 1,
          "ma": "KH001",
          "ten": "Công ty TNHH ABC",
          "diachi": "123 Đường ABC, Quận 1, TP.HCM",
          "dienthoai": "0123456789"
        }
      },
      "KhachhangCreate": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "ma": {
            "type": "string",
            "description": "Mã khách hàng"
          },
          "ten": {
            "type": "string",
            "description": "Tên khách hàng"
          },
          "diachi": {
            "type": "string",
            "description": "Địa chỉ"
          },
          "dienthoai": {
            "type": "string",
            "description": "Số điện thoại"
          }
        },
        "example": {
          "ma": "KH001",
          "ten": "Công ty TNHH ABC",
          "diachi": "123 Đường ABC, Quận 1, TP.HCM",
          "dienthoai": "0123456789"
        }
      },
      "KhachhangUpdate": {
        "type": "object",
        "properties": {
          "ma": {
            "type": "string",
            "description": "Mã khách hàng"
          },
          "ten": {
            "type": "string",
            "description": "Tên khách hàng"
          },
          "diachi": {
            "type": "string",
            "description": "Địa chỉ"
          },
          "dienthoai": {
            "type": "string",
            "description": "Số điện thoại"
          }
        },
        "example": {
          "ten": "Công ty TNHH ABC (Chi nhánh)",
          "diachi": "456 Đường XYZ, Quận 2, TP.HCM",
          "dienthoai": "0987654321"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/khachhang": {
      "get": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Lấy danh sách khách hàng",
        "description": "Lấy toàn bộ danh sách khách hàng trong hệ thống",
        "operationId": "getAllKhachhangs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách khách hàng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Danh sách khách hàng"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Khachhang"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Tạo khách hàng mới",
        "description": "Tạo một khách hàng mới trong hệ thống",
        "operationId": "createKhachhang",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KhachhangCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo khách hàng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo khách hàng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Khachhang"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/khachhang/search": {
      "get": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Tìm kiếm khách hàng theo tên",
        "description": "Tìm kiếm khách hàng theo tên hoặc trả về tất cả nếu không có từ khóa",
        "operationId": "searchKhachhangByName",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Từ khóa tìm kiếm tên khách hàng"
          }
        ],
        "responses": {
          "200": {
            "description": "Kết quả tìm kiếm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Kết quả tìm kiếm"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Khachhang"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/khachhang/{id}": {
      "get": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Lấy thông tin khách hàng theo ID",
        "description": "Lấy chi tiết một khách hàng dựa trên ID",
        "operationId": "getKhachhangById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết khách hàng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết khách hàng"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Khachhang"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Cập nhật khách hàng",
        "description": "Cập nhật thông tin khách hàng dựa trên ID",
        "operationId": "updateKhachhang",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KhachhangUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật khách hàng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật khách hàng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Khachhang"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Xóa khách hàng",
        "description": "Xóa một khách hàng dựa trên ID",
        "operationId": "deleteKhachhang",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa khách hàng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa khách hàng thành công"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/khachhang/code/{ma}": {
      "get": {
        "tags": [
          "Khách Hàng"
        ],
        "summary": "Lấy thông tin khách hàng theo mã",
        "description": "Lấy chi tiết một khách hàng dựa trên mã",
        "operationId": "getKhachhangByMa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ma",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Mã khách hàng"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết khách hàng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chi tiết khách hàng"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Khachhang"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}